requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
SQLAlchemy==1.4.40
subnet-utils==0.5.1
urllib3==1.26.12
yarl==1.8.1
